rename pdb_structure to structure and Structure to PDBStructure
