+++ tests/features/test_atom.py
@@ -2,7 +2,7 @@ import numpy as np
 from pdb2sql import pdb2sql
 from deeprankcore.features.atom import add_features
 from deeprankcore.operations.graph import build_atomic_graph
-from deeprankcore.molstruct.pdb_structure import Chain
+from deeprankcore.molstruct.structure import Chain
 from deeprankcore.molstruct.residue import Residue
 from deeprankcore.operations.buildgraph import get_structure, get_surrounding_residues
 from deeprankcore.domain import nodefeatures
